+++ src/lib/classnames/index.ts
@@ -1,6 +1,6 @@
-import { twMerge } from 'tailwind-merge';
-
 import { type ClassValue, clsx } from 'clsx';
+
+import { twMerge } from 'tailwind-merge';
 
 export default function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
